+++ lua/tmux/resize.lua
@@ -1,10 +1,11 @@
+local layout = require("tmux.layout")
 local keymaps = require("tmux.keymaps")
 local nvim = require("tmux.wrapper.nvim")
 local options = require("tmux.configuration.options")
 local tmux = require("tmux.wrapper.tmux")
-local utils = require("tmux.utils")
 
 local M = {}
+
 function M.setup()
     if options.resize.enable_default_keybindings then
         keymaps.register("n", {
@@ -19,7 +20,7 @@ end
 function M.to_left(step)
     step = step or options.resize.resize_step_x
     local is_border = nvim.is_nvim_border("l")
-    if (nvim.is_nvim_float() or is_border) and utils.is_tmux_target("l") then
+    if (nvim.is_nvim_float() or is_border) and layout.is_tmux_target("l") then
         tmux.resize("h", step)
     elseif is_border then
         nvim.resize("x", "+", step)
@@ -31,7 +32,7 @@ end
 function M.to_bottom(step)
     step = step or options.resize.resize_step_y
     local is_border = nvim.is_nvim_border("j")
-    if (nvim.is_nvim_float() or is_border) and utils.is_tmux_target("j") then
+    if (nvim.is_nvim_float() or is_border) and layout.is_tmux_target("j") then
         tmux.resize("j", step)
     elseif is_border then
         nvim.resize("y", "-", step)
@@ -43,7 +44,7 @@ end
 function M.to_top(step)
     step = step or options.resize.resize_step_y
     local is_border = nvim.is_nvim_border("j")
-    if (nvim.is_nvim_float() or is_border) and utils.is_tmux_target("j") then
+    if (nvim.is_nvim_float() or is_border) and layout.is_tmux_target("j") then
         tmux.resize("k", step)
     elseif is_border then
         nvim.resize("y", "+", step)
@@ -55,7 +56,7 @@ end
 function M.to_right(step)
     step = step or options.resize.resize_step_x
     local is_border = nvim.is_nvim_border("l")
-    if (nvim.is_nvim_float() or is_border) and utils.is_tmux_target("l") then
+    if (nvim.is_nvim_float() or is_border) and layout.is_tmux_target("l") then
         tmux.resize("l", step)
     elseif is_border then
         nvim.resize("x", "-", step)
